Fix / update SCM meta-data in POM
The "connection" URL was missing the generic "git" username, but it
should be a read-only / anonymous clone URL anyway, so just use the
HTTPS URl GitHub provides for this case. The "developerConnection" URL
is only updated to the new project name on GitHub. Finally, the
browsable "url" is updated and prefix with the HTTPS protocol.