Add new option for further customization of collections
In order to support Collections that are quite different than the ones
in the standard library, we add a new `collection` option for repeated
fields. The collection can reference an `Adapter` type available at
runtime that will forward calls for an underlying collection.

This helps support types such as cats.data.NonEmptyList, NonEmptySet
and NonEmptyMap. For example, those collections don't define `foreach`,
and `++` where the rhs is an iterable. For `NonEmptySet` there is even
no `size` method.

See #1013 and scalapb/scalapb-validate#38