Add HasBuilder, improves binary compatibility between 0.10.x and 0.11.x
On the 0.10.x branch, 9a0561901eb7d16827a837f7d22b283bd3cf9daf
introduced a `HasBuilder` trait.

This trait was not introduced to the 0.11.x branch. Because of this,
code generated with 0.10.x of the generator cannot be ran with
0.11.x of the runtime.

Of course we cannot expect 0.10.x to be fully binary compatible with
0.11.x . However, some testing suggests that just making this class
and the `MessageBuilder` interface available on the classpath allows
using the 0.10.x-generated classes with the 0.11.x runtime in simple
cases - allowing for a smoother upgrade path to 0.11.x.

We should stronly encourage people to move to 0.11.x, but having a
ScalaPB release that supports this gives a smoother upgrade path.

Fixes #1153